requirements.txt
openai
gradio
python-dotenv
fastapi
uvicorn